Fixed merge changes
