+++ src/main.rs
@@ -42,6 +42,7 @@
 //!
 //! Contributions to improve `era-test-node` are welcome. Please refer to the contribution guidelines for more details.
 
+use crate::node::{ShowStorageLogs, ShowVMDetails};
 use clap::{Parser, Subcommand};
 use configuration_api::ConfigurationApiNamespaceT;
 use fork::ForkDetails;
@@ -192,86 +193,6 @@ struct Cli {
     #[arg(long)]
     /// If true, will load the locally compiled system contracts (useful when doing changes to system contracts or bootloader)
     dev_use_local_contracts: bool,
-}
-
-#[derive(Debug, Parser, Clone, clap::ValueEnum, PartialEq, Eq)]
-pub enum ShowCalls {
-    None,
-    User,
-    System,
-    All,
-}
-
-impl FromStr for ShowCalls {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_ref() {
-            "none" => Ok(ShowCalls::None),
-            "user" => Ok(ShowCalls::User),
-            "system" => Ok(ShowCalls::System),
-            "all" => Ok(ShowCalls::All),
-            _ => Err(()),
-        }
-    }
-}
-
-impl Display for ShowCalls {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        write!(f, "{:?}", self)
-    }
-}
-
-#[derive(Debug, Parser, Clone, clap::ValueEnum, PartialEq, Eq)]
-pub enum ShowStorageLogs {
-    None,
-    Read,
-    Write,
-    All,
-}
-
-impl FromStr for ShowStorageLogs {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_ref() {
-            "none" => Ok(ShowStorageLogs::None),
-            "read" => Ok(ShowStorageLogs::Read),
-            "write" => Ok(ShowStorageLogs::Write),
-            "all" => Ok(ShowStorageLogs::All),
-            _ => Err(()),
-        }
-    }
-}
-
-impl Display for ShowStorageLogs {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        write!(f, "{:?}", self)
-    }
-}
-
-#[derive(Debug, Parser, Clone, clap::ValueEnum, PartialEq, Eq)]
-pub enum ShowVMDetails {
-    None,
-    All,
-}
-
-impl FromStr for ShowVMDetails {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_ref() {
-            "none" => Ok(ShowVMDetails::None),
-            "all" => Ok(ShowVMDetails::All),
-            _ => Err(()),
-        }
-    }
-}
-
-impl Display for ShowVMDetails {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        write!(f, "{:?}", self)
-    }
 }
 
 #[derive(Debug, Subcommand)]
