+++ src/node.rs
@@ -6,17 +6,20 @@ use crate::{
     formatter,
     formatter::ShowCalls,
     utils::{adjust_l1_gas_price_for_tx, derive_gas_estimation_overhead, IntoBoxedFuture},
-    ShowStorageLogs, ShowVMDetails,
 };
+use clap::Parser;
 use colored::Colorize;
+use core::fmt::Display;
 use futures::FutureExt;
 use jsonrpc_core::BoxFuture;
 use std::{
     cmp::{self},
     collections::HashMap,
     convert::TryInto,
+    str::FromStr,
     sync::{Arc, RwLock},
 };
+
 use vm::{
     utils::{BLOCK_GAS_LIMIT, ETH_CALL_GAS_LIMIT},
     vm::VmTxExecutionResult,
@@ -97,6 +100,58 @@ pub struct TxExecutionInfo {
     pub batch_number: u32,
     pub miniblock_number: u64,
     pub result: VmTxExecutionResult,
+}
+
+#[derive(Debug, Parser, Clone, clap::ValueEnum, PartialEq, Eq)]
+pub enum ShowStorageLogs {
+    None,
+    Read,
+    Write,
+    All,
+}
+
+impl FromStr for ShowStorageLogs {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_ref() {
+            "none" => Ok(ShowStorageLogs::None),
+            "read" => Ok(ShowStorageLogs::Read),
+            "write" => Ok(ShowStorageLogs::Write),
+            "all" => Ok(ShowStorageLogs::All),
+            _ => Err(()),
+        }
+    }
+}
+
+impl Display for ShowStorageLogs {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[derive(Debug, Parser, Clone, clap::ValueEnum, PartialEq, Eq)]
+pub enum ShowVMDetails {
+    None,
+    All,
+}
+
+impl FromStr for ShowVMDetails {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_ref() {
+            "none" => Ok(ShowVMDetails::None),
+            "all" => Ok(ShowVMDetails::All),
+            _ => Err(()),
+        }
+    }
+}
+
+impl Display for ShowVMDetails {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "{:?}", self)
+    }
 }
 
 /// Helper struct for InMemoryNode.
@@ -547,9 +602,9 @@ impl Default for InMemoryNode {
     fn default() -> Self {
         InMemoryNode::new(
             None,
-            crate::ShowCalls::None,
-            crate::ShowStorageLogs::None,
-            crate::ShowVMDetails::None,
+            crate::formatter::ShowCalls::None,
+            ShowStorageLogs::None,
+            ShowVMDetails::None,
             false,
             false,
         )
